services/payment-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
python-json-logger==2.0.7
